(DIO-409) Add custom post-provisioning scripting
Prior to this commit there was no way to override what scripts were run
as the last part of the provisioning process.
This commit updates the previous last shell to use an array of scripts
coming from `post_provisioning_scripts`. It also allows customizing what
exit codes are accepted from the passed in scripts.

Additionally, as an example of using the scripts, I have added the
PuppetDB pre-loading scripts that Nick Burgan has been working on
and plan to use those as a test that I have successfully made the new
customizable "role" feature that I've added.