(DIO-409) Add Rakefile
Prior to this commit there was no automated way to perform the packer
builds. Instead, users had to copypaste a smattering of commands out
of README.md files.
This commit adds a `Rakefile` that contains a few tasks that mirror the
existing `Makefile` as well as add a new `build` task that will build a
template based on several ENV variables that can be set. Documentation
on those variables will be added as I go along.